requirements.txt
appdirs==1.4.4
asgiref==3.3.4
asn1crypto==0.24.0
astroid==1.6.6
autopep8==1.5.6
Babel==2.9.1
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
black==20.8b1
blinker==1.4
certifi==2019.9.11
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==4.0.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==2.8
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-factory-generator==1.0.4
django-filter==2.4.0
django-guardian==2.3.0
django-phonenumber-field==5.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
keyring==18.0.1
keyrings.alt==3.4.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.9
paramiko==2.6.0
pathspec==0.8.1
pbr==5.6.0
pexpect==4.6.0
phonenumbers==8.12.22
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.7.3
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.26
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.6
ruamel.yaml.clib==0.2.2
SecretStorage==2.3.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
testresources==2.0.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==20.0.17
wadllib==1.3.3
whitenoise==4.1.3
wrapt==1.12.1
xlrd==1.2.0
youtube-dl==2020.3.24
zipp==1.0.0

